feat: add @reduxjs/toolkit and react-redux dependencies
The project now requires the use of Redux for state management. The "@reduxjs/toolkit" package has been added as a dependency to provide a set of tools and utilities for working with Redux. The "react-redux" package has also been added as a dependency to integrate Redux with React components.

feat: create ActionsButton component

A new component called ActionsButton has been created in the src/components/redux/Actions directory. This component uses the useDispatch hook from react-redux to dispatch the increment and decrement actions from the counterSlice feature. It renders two buttons, one for incrementing the count and one for decrementing the count.

feat: add counterSlice feature

A new feature called counterSlice has been added in the src/features directory. This feature defines the initial state of the count as 0 and provides two reducers: increment and decrement. The increment reducer increases the count by 1, while the decrement reducer decreases the count by 1 if it is greater than 1.

fix: import Provider and store in main.tsx

The Provider component from react-redux and the store from the src/store directory have been imported in the src/main.tsx file. The Provider component wraps the App component to provide access to the Redux store throughout the application.

feat: update Home page to use counterSlice

The Home component in the src/pages/home directory has been updated to use the counterSlice feature. The useSelector hook from react-redux is used to select the count from the Redux store. The count is then rendered in a paragraph element below the heading. The ActionsButton component is also rendered to allow the user to increment or decrement the count.

feat: create store configuration

A new file called index.ts has been created in the src/store directory. This file exports a configured Redux store using the configureStore function from @reduxjs/toolkit. The counterReducer from the counterSlice feature is passed as the reducer to the store configuration. The RootState type is also exported to provide type safety when accessing the store state.